requirements.txt
streamlit>=1.29.0
langchain==0.1.17
langchain-google-genai==0.0.10
python-dotenv>=1.0.0
